Move save core case service out of ccd-client
