--- ccd-client/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=uk.gov.hmcts.cmc.ccd.client.SaveCaseService
